+++ src/library/adash-messages.adb
@@ -543,6 +543,8 @@ package body Adash.Messages is
          when Msg_Command_Run_All_New_Doc =>
             return "command.run_all_new.doc";
          when Msg_Command_Pipe_Doc           => return "command.pipe.doc";
+         when Msg_Command_Pipe_Start_Doc =>
+            return "command.pipe_start.doc";
          when Msg_Command_Pipe_From_Doc =>
             return "command.pipe_from.doc";
          when Msg_Command_Pipe_Into_Doc =>
@@ -918,6 +920,7 @@ package body Adash.Messages is
             | Msg_Command_Run_All_Into_Doc | Msg_Command_Run_All_Append_Doc
             | Msg_Command_Run_All_New_Doc
             | Msg_Command_Pipe_Doc
+            | Msg_Command_Pipe_Start_Doc
             | Msg_Command_Pipe_From_Doc
             | Msg_Command_Pipe_Into_Doc
             | Msg_Command_Pipe_Append_Doc